backend/requirements.txt
apturl==0.5.2
asgiref==3.5.0
backports.zoneinfo==0.2.1
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.4
distro==1.4.0
distro-info===0.23ubuntu1
Django==4.0.2
django-cors-headers==3.11.0
djangorestframework==3.13.1
entrypoints==0.3
filelock==3.6.0
httplib2==0.14.0
idna==2.8
keyring==18.0.1


virtualenv==20.13.2
virtualenv-clone==0.5.7

